BackEnd/requirements.txt
aliyun-python-sdk-core==2.13.9
aliyun-python-sdk-core-v3==2.13.9
aliyun-python-sdk-kms==2.7.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
crcmod==1.7
Django==2.2.4
django-cors-headers==3.1.0
django-crontab==0.7.1
django-prometheus==1.0.15
djangorestframework==3.10.2
idna==2.8
itsdangerous==1.1.0
jmespath==0.9.4
loguru==0.3.2
mongoengine==0.18.2
oss2==2.8.0
prometheus-client==0.7.1
pycryptodome==3.9.0
pymongo==3.6.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
win32-setctime==1.0.1
